Feature 22/select change benefit (#37)
Main Changes
Added the EmployeeBenefits.vue component.
Integrated an API call (GET /api/benefit) to fetch available benefits.
Implemented a modal interface for selecting benefits.
Added real-time validation to ensure users cannot select more than the allowed number of benefits (maxBenefits).
Included functionality to remove selected benefits.
